added Authentication to User- and OrderController

+++ app/Http/Controllers/OrderController.php
@@ -12,13 +12,18 @@ use Carbon\Carbon;
 class OrderController extends Controller
 {
     //index orders
-    public function index(){
-        $orders = Order::with(['owner' => function ($query){
-            $query->select('id', 'name');
-        }])->get();
-
-//        return $orders;
-        return view('orders/Orders', compact('orders'));
+    public function index(Request $req){
+        if(in_array("read", $req->get('permissions'))){
+            $orders = Order::with(['owner' => function ($query){
+                $query->select('id', 'name');
+            }])->get();
+    
+            //return $orders;
+            return view('orders/Orders', compact('orders'));
+        }
+        else {
+            return abort(403, "Sorry you do not have the right permissions");
+        }
     }
 
     //show
@@ -64,44 +69,55 @@ class OrderController extends Controller
     }
 
     //edit
-    public function edit($id){
-        $order = Order::findOrFail($id);
+    public function edit($id, Request $req){
+        if (in_array("write", $req->get('permissions'))) {
+            $order = Order::findOrFail($id);
 
-        return view('orders.edit', [
-            'order' => $order,
-        ]);
+            return view('orders.edit', [
+                'order' => $order,
+            ]);
+        } else {
+            abort(403, "Sorry you do not have the right permissions");
+        }
     }
 
     //update
     public function update(Request $req ,$id){
-
-        //find corresponding order.
-        $order = Order::findOrFail($id);
-
-        //update data
-        $order->clip = $req->clip;
-        $order->engraving = $req->engraving;
-        $order->front_img = $req->front_img;
-        $order->back_img = $req->back_img;
-        $order->location = $req->location;
-        $order->cup_id = $req->cup_id;
-        $order->status = $req->status;
-        $order->user_id = $req->user_id;
-        
-        $order->save();
-
-        return redirect()->route('orders.show', ['id' => $id]);
+        if (in_array("write", $req->get('permissions'))){
+            //find corresponding order.
+            $order = Order::findOrFail($id);
+    
+            //update data
+            $order->clip = $req->clip;
+            $order->engraving = $req->engraving;
+            $order->front_img = $req->front_img;
+            $order->back_img = $req->back_img;
+            $order->location = $req->location;
+            $order->cup_id = $req->cup_id;
+            $order->status = $req->status;
+            $order->user_id = $req->user_id;
+            
+            $order->save();
+    
+            return redirect()->route('orders.show', ['id' => $id]);
+        } else {
+            return abort(403, "Sorry you do not have the right permissions");
+        }
     }
 
     //delete
-    public function delete($id){
-        $deletedOrder = Order::destroy($id);
-
-        if ($deletedOrder){
-            return redirect()->route('orders.index');
-        }
-        else{
-            return "Oops something went wrong";
+    public function delete($id, Request $req){
+        if (in_array('delete', $req->get('permissions'))) {
+            $deletedOrder = Order::destroy($id);
+    
+            if ($deletedOrder){
+                return redirect()->route('orders.index');
+            }
+            else{
+                return "Oops something went wrong";
+            }
+        } else {
+            abort(403, "Sorry you do not have the right permissions");
         }
     }
 }
